fix: address review — keep backend private by co-locating AdMob extension
