+++ Sources/Purchasing/Purchases/Purchases.swift
@@ -273,7 +273,7 @@ public typealias StartPurchaseBlock = (@escaping PurchaseCompletedBlock) -> Void
 
     private let attributionFetcher: AttributionFetcher
     private let attributionPoster: AttributionPoster
-    let backend: Backend
+    private let backend: Backend
     private let deviceCache: DeviceCache
     private let paywallCache: PaywallCacheWarmingType?
     private let identityManager: IdentityManager
@@ -1566,6 +1566,43 @@ public extension Purchases {
     /// Used by `RevenueCatUI` to download and cache paywall images.
     @available(iOS 15.0, macOS 12.0, watchOS 8.0, tvOS 15.0, *)
     static let paywallImageDownloadSession: URLSession = PaywallCacheWarming.downloadSession
+
+}
+
+// MARK: - AdMob SSV (Internal SPI)
+
+extension Purchases {
+
+    /// Polls the backend once for AdMob SSV verification status using `client_transaction_id`.
+    ///
+    /// Internal API for RC ad adapters.
+    ///
+    /// Cancelling the calling `Task` does not cancel the in-flight HTTP request.
+    @_spi(Internal) public func pollAdMobSSVStatus(
+        clientTransactionID: String
+    ) async throws -> AdMobSSVPollStatus {
+        let response = try await Async.call { completion in
+            self.backend.adsAPI.getAdMobSSVStatus(
+                appUserID: self.appUserID,
+                clientTransactionID: clientTransactionID
+            ) { result in
+                completion(result.mapError(\.asPublicError))
+            }
+        }
+
+        switch response.status {
+        case .validated:
+            return .validated
+        case .pending:
+            return .pending
+        case .failed:
+            return .failed
+        case .unknown:
+            // Defensive: treat unrecognized future statuses as still-pending so the
+            // adapter keeps polling rather than firing a false terminal outcome.
+            return .pending
+        }
+    }
 
 }
